docker is working

+++ src/src/Scanner.php
@@ -14,7 +14,16 @@ require __DIR__ . '/../vendor/autoload.php';
 require_once('Dal.php');
 require_once('Media.php');
 
+
+Dal::waitUntilOnline(getenv('MONGO_URL'));
+
 $dal = new Dal();
+// $dal->waitUntilOnline();
+if ($dal->mediaCollectionExists()){
+    printf("Db exists already don't know how to update....");
+    exit();
+}
+
 if (array_key_exists("dropdb", $options)  && !$options["dropdb"]) {
     print "Dropping db\r\n";
     $dal->drop();
@@ -23,12 +32,13 @@ if (array_key_exists("dropdb", $options)  && !$options["dropdb"]) {
 $baseDir = getenv('BASEDIR');
 $thumbDir = getenv('THUMBDIR');
 
-
 if ($handle = opendir($baseDir)) {
-    $root[] = Media::withAbsoluteDirAndFilename($baseDir, '/');
-    $dal->insertRecords($root);
+    // $root[] = Media::withAbsoluteDirAndFilename($baseDir, '/');
+    // $dal->insertRecords($root);
     getChildren($baseDir, $dal);
 }
+
+closedir($handle);
 
 function getChildren($dir, $dal)
 {
